Get bulk fingerprint route refactored
