+++ modules/controller/dv-controller.js
@@ -903,76 +903,6 @@ class DVController {
         });
     }
 
-    async handleGetFingerprint(req, res) {
-        this.logger.api('GET: Fingerprint request received.');
-        const { dataset_id } = req.params;
-        if (!dataset_id) {
-            res.status(400);
-            res.send({
-                message: 'data_set_id parameter is missing',
-            });
-            return;
-        }
-        try {
-            const { result, replicated } = await this.dvService.getFingerprintData(dataset_id);
-
-            if (replicated) {
-                res.status(200);
-            } else {
-                res.status(404);
-            }
-            res.send(result);
-        } catch (e) {
-            const message = `Failed to get fingerprints. ${e.message}.`;
-            this.logger.error(message);
-            res.status(400);
-            res.send({
-                message,
-            });
-        }
-    }
-
-
-    async handleGetBulkFingerprint(req, res) {
-        this.logger.api('GET: Bulk fingerprint bulk request received.');
-        if (!req.body || !req.body.dataset_ids) {
-            res.status(400);
-            res.send({
-                message: 'dataset_ids parameter is empty',
-            });
-            return;
-        }
-
-        try {
-            const inserted_object = await Models.handler_ids.create({
-                status: 'PENDING',
-            });
-
-            const { handler_id } = inserted_object.dataValues;
-
-            await this.commandExecutor.add({
-                name: 'dvFingerprintCommand',
-                sequence: [],
-                delay: 0,
-                data: {
-                    handler_id,
-                    dataset_ids: req.body.dataset_ids,
-                },
-                transactional: false,
-            });
-
-            res.status(200);
-            res.send({
-                handler_id,
-            });
-        } catch (e) {
-            res.status(400);
-            res.send({
-                message: e.message,
-            });
-        }
-    }
-
     /**
      * Handle new purchase on the blockchain and add the node that bought data as a new
      * data seller
